requirements.txt
faiss-cpu==1.10.0
flask==3.1.0
flask-restful==0.3.10
numpy==2.2.3
pytest==8.3.4
scikit-learn==1.6.1